Delete EntityLog and migrate system events from `EntityLogId` and `DispatchedMessageId` to `MessageId`
