+++ server/src/main/java/io/spine/server/entity/EntityLifecycle.java
@@ -25,18 +25,17 @@ import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Any;
 import io.spine.annotation.Internal;
 import io.spine.base.EventMessage;
+import io.spine.base.Identifier;
+import io.spine.client.EntityId;
 import io.spine.core.Command;
 import io.spine.core.CommandId;
 import io.spine.core.Event;
 import io.spine.core.EventId;
 import io.spine.core.MessageId;
-import io.spine.core.SignalId;
 import io.spine.core.Version;
 import io.spine.option.EntityOption;
 import io.spine.system.server.CommandTarget;
 import io.spine.system.server.ConstraintViolated;
-import io.spine.system.server.DispatchedMessageId;
-import io.spine.system.server.EntityLogId;
 import io.spine.system.server.SystemWriteSide;
 import io.spine.system.server.event.CommandDispatchedToHandler;
 import io.spine.system.server.event.CommandHandled;
@@ -62,8 +61,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static io.spine.base.Time.currentTime;
 import static io.spine.server.entity.EventFilter.allowAll;
-import static io.spine.util.Exceptions.newIllegalArgumentException;
-import static java.util.stream.Collectors.toList;
 
 /**
  * The lifecycle callbacks of an {@link Entity}.
@@ -90,14 +87,13 @@ public class EntityLifecycle {
     private final EventFilter eventFilter;
 
     /**
-     * The ID of {@linkplain io.spine.system.server.EntityLog history} of the associated
-     * {@link Entity}.
+     * The message ID of of the associated {@link Entity} state.
      *
      * <p>Most commands posted by the {@code EntityLifecycle} are handled by
      * the {@code io.spine.system.server.EntityHistoryAggregate}.
      * Thus, storing an ID as a field is convenient.
      */
-    private final EntityLogId historyId;
+    private final MessageId entityId;
 
     /**
      * Creates a new instance.
@@ -113,11 +109,18 @@ public class EntityLifecycle {
                               EventFilter eventFilter) {
         this.systemWriteSide = checkNotNull(writeSide);
         this.eventFilter = checkNotNull(eventFilter);
-        this.historyId = EntityLogIds.wrap(entityId, entityType);
+        this.entityId = MessageId
+                .newBuilder()
+                .setId(Identifier.pack(entityId))
+                .setTypeUrl(entityType.value())
+                .vBuild();
     }
 
     private EntityLifecycle(Builder builder) {
-        this(builder.entityId, builder.entityType, builder.writeSide, builder.eventFilter);
+        this(builder.entityId,
+             builder.entityType,
+             builder.writeSide,
+             builder.eventFilter);
     }
 
     /**
@@ -129,7 +132,7 @@ public class EntityLifecycle {
     public final void onEntityCreated(EntityOption.Kind entityKind) {
         EntityCreated event = EntityCreated
                 .newBuilder()
-                .setId(historyId)
+                .setEntity(entityId)
                 .setKind(entityKind)
                 .vBuild();
         postEvent(event);
@@ -143,10 +146,14 @@ public class EntityLifecycle {
      *         the ID of the command which should be handled by the entity
      */
     public final void onTargetAssignedToCommand(CommandId commandId) {
+        EntityId entityId = EntityId
+                .newBuilder()
+                .setId(this.entityId.getId())
+                .buildPartial();
         CommandTarget target = CommandTarget
                 .newBuilder()
-                .setEntityId(historyId.getEntityId())
-                .setTypeUrl(historyId.getTypeUrl())
+                .setEntityId(entityId)
+                .setTypeUrl(this.entityId.getTypeUrl())
                 .vBuild();
         TargetAssignedToCommand event = TargetAssignedToCommand
                 .newBuilder()
@@ -165,7 +172,7 @@ public class EntityLifecycle {
     public final void onDispatchCommand(Command command) {
         CommandDispatchedToHandler systemCommand = CommandDispatchedToHandler
                 .newBuilder()
-                .setReceiver(historyId)
+                .setReceiver(entityId)
                 .setPayload(command)
                 .setWhenDispatched(currentTime())
                 .vBuild();
@@ -212,7 +219,7 @@ public class EntityLifecycle {
     public final void onDispatchEventToSubscriber(Event event) {
         EventDispatchedToSubscriber systemCommand = EventDispatchedToSubscriber
                 .newBuilder()
-                .setReceiver(historyId)
+                .setReceiver(entityId)
                 .setPayload(event)
                 .setWhenDispatched(currentTime())
                 .vBuild();
@@ -222,7 +229,7 @@ public class EntityLifecycle {
     public final void onEventImported(Event event) {
         EventImported systemEvent = EventImported
                 .newBuilder()
-                .setReceiver(historyId)
+                .setReceiver(entityId)
                 .setEventId(event.getId())
                 .setWhenImported(currentTime())
                 .vBuild();
@@ -238,7 +245,7 @@ public class EntityLifecycle {
     public final void onDispatchEventToReactor(Event event) {
         EventDispatchedToReactor systemCommand = EventDispatchedToReactor
                 .newBuilder()
-                .setReceiver(historyId)
+                .setReceiver(entityId)
                 .setPayload(event)
                 .setWhenDispatched(currentTime())
                 .vBuild();
@@ -259,14 +266,12 @@ public class EntityLifecycle {
      *         {@link EventId EventId}s or {@link CommandId}s
      */
     public final void onStateChanged(EntityRecordChange change,
-                                     Set<? extends SignalId> messageIds) {
-        Collection<DispatchedMessageId> dispatchedMessageIds = toDispatched(messageIds);
-
-        postIfChanged(change, dispatchedMessageIds);
-        postIfArchived(change, dispatchedMessageIds);
-        postIfDeleted(change, dispatchedMessageIds);
-        postIfExtracted(change, dispatchedMessageIds);
-        postIfRestored(change, dispatchedMessageIds);
+                                     Set<? extends MessageId> messageIds) {
+        postIfChanged(change, messageIds);
+        postIfArchived(change, messageIds);
+        postIfDeleted(change, messageIds);
+        postIfExtracted(change, messageIds);
+        postIfRestored(change, messageIds);
     }
 
     /**
@@ -287,8 +292,8 @@ public class EntityLifecycle {
                                       Version version) {
         MessageId entityId = MessageId
                 .newBuilder()
-                .setId(historyId.getEntityId().getId())
-                .setTypeUrl(historyId.getTypeUrl())
+                .setId(this.entityId.getId())
+                .setTypeUrl(this.entityId.getTypeUrl())
                 .setVersion(version)
                 .buildPartial();
         ConstraintViolated event = ConstraintViolated
@@ -302,7 +307,7 @@ public class EntityLifecycle {
     }
 
     private void postIfChanged(EntityRecordChange change,
-                               Collection<DispatchedMessageId> messageIds) {
+                               Collection<? extends MessageId> messageIds) {
         Any oldState = change.getPreviousValue()
                              .getState();
         Any newState = change.getNewValue()
@@ -312,9 +317,9 @@ public class EntityLifecycle {
                                        .getVersion();
             EntityStateChanged event = EntityStateChanged
                     .newBuilder()
-                    .setId(historyId)
+                    .setEntity(entityId)
                     .setNewState(newState)
-                    .addAllMessageId(ImmutableList.copyOf(messageIds))
+                    .addAllSignalId(ImmutableList.copyOf(messageIds))
                     .setNewVersion(newVersion)
                     .vBuild();
             postEvent(event);
@@ -322,7 +327,7 @@ public class EntityLifecycle {
     }
 
     private void postIfArchived(EntityRecordChange change,
-                                Collection<DispatchedMessageId> messageIds) {
+                                Collection<? extends MessageId> messageIds) {
         boolean oldValue = change.getPreviousValue()
                                  .getLifecycleFlags()
                                  .getArchived();
@@ -334,8 +339,8 @@ public class EntityLifecycle {
                                     .getVersion();
             EntityArchived event = EntityArchived
                     .newBuilder()
-                    .setId(historyId)
-                    .addAllMessageId(ImmutableList.copyOf(messageIds))
+                    .setEntity(entityId)
+                    .addAllSignalId(ImmutableList.copyOf(messageIds))
                     .setVersion(version)
                     .vBuild();
             postEvent(event);
@@ -343,7 +348,7 @@ public class EntityLifecycle {
     }
 
     private void postIfDeleted(EntityRecordChange change,
-                               Collection<DispatchedMessageId> messageIds) {
+                               Collection<? extends MessageId> messageIds) {
         boolean oldValue = change.getPreviousValue()
                                  .getLifecycleFlags()
                                  .getDeleted();
@@ -355,8 +360,8 @@ public class EntityLifecycle {
                                     .getVersion();
             EntityDeleted event = EntityDeleted
                     .newBuilder()
-                    .setId(historyId)
-                    .addAllMessageId(ImmutableList.copyOf(messageIds))
+                    .setEntity(entityId)
+                    .addAllSignalId(ImmutableList.copyOf(messageIds))
                     .setVersion(version)
                     .vBuild();
             postEvent(event);
@@ -364,7 +369,7 @@ public class EntityLifecycle {
     }
 
     private void postIfExtracted(EntityRecordChange change,
-                                 Collection<DispatchedMessageId> messageIds) {
+                                 Collection<? extends MessageId> messageIds) {
         boolean oldValue = change.getPreviousValue()
                                  .getLifecycleFlags()
                                  .getArchived();
@@ -376,8 +381,8 @@ public class EntityLifecycle {
                                     .getVersion();
             EntityUnarchived event = EntityUnarchived
                     .newBuilder()
-                    .setId(historyId)
-                    .addAllMessageId(ImmutableList.copyOf(messageIds))
+                    .setEntity(entityId)
+                    .addAllSignalId(ImmutableList.copyOf(messageIds))
                     .setVersion(version)
                     .vBuild();
             postEvent(event);
@@ -385,7 +390,7 @@ public class EntityLifecycle {
     }
 
     private void postIfRestored(EntityRecordChange change,
-                                Collection<DispatchedMessageId> messageIds) {
+                                Collection<? extends MessageId> messageIds) {
         boolean oldValue = change.getPreviousValue()
                                  .getLifecycleFlags()
                                  .getDeleted();
@@ -397,8 +402,8 @@ public class EntityLifecycle {
                                     .getVersion();
             EntityRestored event = EntityRestored
                     .newBuilder()
-                    .setId(historyId)
-                    .addAllMessageId(ImmutableList.copyOf(messageIds))
+                    .setEntity(entityId)
+                    .addAllSignalId(ImmutableList.copyOf(messageIds))
                     .setVersion(version)
                     .vBuild();
             postEvent(event);
@@ -408,35 +413,6 @@ public class EntityLifecycle {
     protected void postEvent(EventMessage event) {
         Optional<? extends EventMessage> filtered = eventFilter.filter(event);
         filtered.ifPresent(systemWriteSide::postEvent);
-    }
-
-    private static Collection<DispatchedMessageId>
-    toDispatched(Collection<? extends SignalId> messageIds) {
-        Collection<DispatchedMessageId> dispatchedMessageIds =
-                messageIds.stream()
-                          .map(EntityLifecycle::dispatchedMessageId)
-                          .collect(toList());
-        return dispatchedMessageIds;
-    }
-
-    @SuppressWarnings("ChainOfInstanceofChecks")
-    private static DispatchedMessageId dispatchedMessageId(SignalId signalId) {
-        checkNotNull(signalId);
-        DispatchedMessageId.Builder builder = DispatchedMessageId.newBuilder();
-        if (signalId instanceof EventId) {
-            EventId eventId = (EventId) signalId;
-            return builder.setEventId(eventId)
-                          .vBuild();
-        } else if (signalId instanceof CommandId) {
-            CommandId commandId = (CommandId) signalId;
-            return builder.setCommandId(commandId)
-                          .vBuild();
-        } else {
-            throw newIllegalArgumentException(
-                    "Unexpected message ID of type %s. Expected EventId or CommandId.",
-                    signalId.getClass()
-            );
-        }
     }
 
     /**
