+++ server/src/main/java/io/spine/server/stand/Stand.java
@@ -25,6 +25,7 @@ import com.google.common.util.concurrent.MoreExecutors;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import com.google.errorprone.annotations.CheckReturnValue;
 import com.google.protobuf.Any;
+import com.google.protobuf.Empty;
 import io.grpc.stub.StreamObserver;
 import io.spine.annotation.Internal;
 import io.spine.base.Identifier;
@@ -34,7 +35,7 @@ import io.spine.client.QueryResponse;
 import io.spine.client.Subscription;
 import io.spine.client.SubscriptionUpdate;
 import io.spine.client.Topic;
-import io.spine.core.EventId;
+import io.spine.core.MessageId;
 import io.spine.core.Response;
 import io.spine.core.Responses;
 import io.spine.core.TenantId;
@@ -98,9 +99,10 @@ public class Stand extends AbstractEventSubscriber implements AutoCloseable {
      *             purposes.
      */
     @Deprecated
-    private static final EventId STAND_POST_ORIGIN = EventId
+    private static final MessageId STAND_POST_ORIGIN = MessageId
             .newBuilder()
-            .setValue("Stand-received-entity-update")
+            .setId(Identifier.pack("Stand-received-entity-update"))
+            .setTypeUrl(TypeUrl.of(Empty.class).value())
             .vBuild();
 
     /**
