--- server/src/main/java/io/spine/system/server/EntityLogProjection.java
@@ -1,178 +1,0 @@
-/*
- * Copyright 2019, TeamDev. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.system.server;
-
-import com.google.protobuf.Timestamp;
-import io.spine.core.Subscribe;
-import io.spine.server.aggregate.Aggregate;
-import io.spine.server.entity.LifecycleFlags;
-import io.spine.server.projection.Projection;
-import io.spine.system.server.event.CommandDispatchedToHandler;
-import io.spine.system.server.event.EntityArchived;
-import io.spine.system.server.event.EntityCreated;
-import io.spine.system.server.event.EntityDeleted;
-import io.spine.system.server.event.EntityRestored;
-import io.spine.system.server.event.EntityStateChanged;
-import io.spine.system.server.event.EntityUnarchived;
-import io.spine.system.server.event.EventDispatchedToReactor;
-import io.spine.system.server.event.EventDispatchedToSubscriber;
-import io.spine.system.server.event.EventImported;
-
-import java.util.function.UnaryOperator;
-
-import static com.google.protobuf.util.Timestamps.compare;
-
-/**
- * The aggregate which manages the history of a single entity.
- *
- * <p>Each {@link Aggregate}, {@link io.spine.server.projection.Projection Projection},
- * and {@link io.spine.server.procman.ProcessManager ProcessManager} in the system has
- * a corresponding entity history.
- *
- * <p>The history of this aggregate has the knowledge about all the messages ever dispatched to
- * the associated entity.
- *
- * <p>An {@code EntityHistory} gives the record-based entities (such as {@code Projection}s and
- * {@code ProcessManager}s) traits of an event-sourced entity. For instance, with the help
- * of {@code EntityHistory} the history of a record-based entity can be investigated and
- * manipulated. The major use case for this facility is implementing idempotent message handlers.
- *
- * <p>This aggregate belongs to the {@code System} bounded context. This aggregate doesn't have
- * an entity history of its own.
- */
-@SuppressWarnings("OverlyCoupledClass") // OK for an entity class.
-final class EntityLogProjection
-        extends Projection<EntityLogId, EntityLog, EntityLog.Builder> {
-
-    @Subscribe
-    void on(EntityCreated event) {
-        builder().setId(event.getId());
-    }
-
-    @Subscribe
-    void on(EventDispatchedToSubscriber event) {
-        builder().setId(event.getReceiver());
-        updateLastEventTime(event.getWhenDispatched());
-    }
-
-    @Subscribe
-    void on(EventDispatchedToReactor event) {
-        builder().setId(event.getReceiver());
-        updateLastEventTime(event.getWhenDispatched());
-    }
-
-    @Subscribe
-    void on(CommandDispatchedToHandler event) {
-        builder().setId(event.getReceiver());
-        updateLastCommandTime(event.getWhenDispatched());
-    }
-
-    @Subscribe
-    void on(EntityStateChanged event) {
-        builder().setId(event.getId())
-                 .setLastStateChange(event.getWhen());
-    }
-
-    @Subscribe
-    void on(EntityArchived event) {
-        updateLifecycleFlags(builder -> builder.setArchived(true));
-        Timestamp whenOccurred = event.getWhen();
-        updateLifecycleTimestamp(builder -> builder.setWhenArchived(whenOccurred));
-    }
-
-    @Subscribe
-    void on(EntityDeleted event) {
-        updateLifecycleFlags(builder -> builder.setDeleted(true));
-        Timestamp whenOccurred = event.getWhen();
-        updateLifecycleTimestamp(builder -> builder.setWhenDeleted(whenOccurred));
-    }
-
-    @Subscribe
-    void on(EntityUnarchived event) {
-        updateLifecycleFlags(builder -> builder.setArchived(false));
-        Timestamp whenOccurred = event.getWhen();
-        updateLifecycleTimestamp(builder -> builder.setWhenExtractedFromArchive(whenOccurred));
-    }
-
-    @Subscribe
-    void on(EntityRestored event) {
-        updateLifecycleFlags(builder -> builder.setDeleted(false));
-        Timestamp whenOccurred = event.getWhen();
-        updateLifecycleTimestamp(builder -> builder.setWhenRestored(whenOccurred));
-    }
-
-    @Subscribe
-    void on(EventImported event) {
-        updateLastEventTime(event.getWhenImported());
-    }
-
-    private void updateLifecycleFlags(UnaryOperator<LifecycleFlags.Builder> mutation) {
-        LifecycleHistory oldLifecycleHistory = builder().getLifecycle();
-        LifecycleFlags.Builder flagsBuilder =
-                oldLifecycleHistory.getLifecycleFlags()
-                                   .toBuilder();
-        LifecycleFlags newFlags = mutation.apply(flagsBuilder)
-                                          .build();
-        LifecycleHistory newLifecycleHistory =
-                oldLifecycleHistory.toBuilder()
-                                   .setLifecycleFlags(newFlags)
-                                   .build();
-        builder().setLifecycle(newLifecycleHistory);
-    }
-
-    private void updateLifecycleTimestamp(UnaryOperator<LifecycleHistory.Builder> mutation) {
-        EntityLog.Builder builder = builder();
-        LifecycleHistory.Builder history =
-                builder.getLifecycle()
-                       .toBuilder();
-        LifecycleHistory newHistory =
-                mutation.apply(history)
-                        .build();
-        builder.setLifecycle(newHistory);
-    }
-
-    private void updateLastEventTime(Timestamp newEvent) {
-        Timestamp lastEvent = builder().getDispatching()
-                                       .getWhenEvent();
-        if (compare(newEvent, lastEvent) > 0) {
-            updateDispatchingHistory(builder -> builder.setWhenEvent(newEvent));
-        }
-    }
-
-    private void updateLastCommandTime(Timestamp newCommand) {
-        Timestamp lastCommand = builder().getDispatching()
-                                         .getWhenCommand();
-        if (compare(newCommand, lastCommand) > 0) {
-            updateDispatchingHistory(builder -> builder.setWhenCommand(newCommand));
-        }
-    }
-
-    private void updateDispatchingHistory(UnaryOperator<DispatchingHistory.Builder> mutation) {
-        EntityLog.Builder builder = builder();
-        DispatchingHistory.Builder history =
-                builder.getDispatching()
-                       .toBuilder();
-        DispatchingHistory newHistory =
-                mutation.apply(history)
-                        .build();
-        builder.setDispatching(newHistory);
-    }
-}
