--- server/src/main/java/io/spine/system/server/EntityLogRepository.java
@@ -1,58 +1,0 @@
-/*
- * Copyright 2019, TeamDev. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.system.server;
-
-import com.google.errorprone.annotations.OverridingMethodsMustInvokeSuper;
-import io.spine.client.EntityId;
-import io.spine.core.MessageId;
-import io.spine.server.projection.ProjectionRepository;
-import io.spine.server.route.EventRoute;
-import io.spine.server.route.EventRouting;
-
-import static io.spine.server.route.EventRoute.withId;
-
-/**
- * The repository for {@link EntityLogProjection}s.
- */
-final class EntityLogRepository
-        extends ProjectionRepository<EntityLogId, EntityLogProjection, EntityLog> {
-
-    @OverridingMethodsMustInvokeSuper
-    @Override
-    protected void setupEventRouting(EventRouting<EntityLogId> routing) {
-        super.setupEventRouting(routing);
-        routing.replaceDefault(EventRoute.byFirstMessageField(EntityLogId.class))
-               .route(ConstraintViolated.class,
-                      (message, context) -> withId(asHistoryId(message.getEntity())));
-    }
-
-    private static EntityLogId asHistoryId(MessageId id) {
-        EntityId entityId = EntityId
-                .newBuilder()
-                .setId(id.getId())
-                .buildPartial();
-        return EntityLogId
-                .newBuilder()
-                .setEntityId(entityId)
-                .setTypeUrl(id.getTypeUrl())
-                .vBuild();
-    }
-}
