+++ server/src/main/java/io/spine/system/server/MirrorRepository.java
@@ -30,6 +30,7 @@ import io.spine.client.Query;
 import io.spine.client.Target;
 import io.spine.client.TargetFilters;
 import io.spine.code.proto.EntityStateOption;
+import io.spine.core.MessageId;
 import io.spine.logging.Logging;
 import io.spine.option.EntityOption;
 import io.spine.option.EntityOption.Kind;
@@ -81,18 +82,18 @@ final class MirrorRepository
     protected void setupEventRouting(EventRouting<MirrorId> routing) {
         super.setupEventRouting(routing);
         routing.route(EntityStateChanged.class,
-                      (message, context) -> targetsFrom(message.getId()))
+                      (message, context) -> targetsFrom(message.getEntity()))
                .route(EntityArchived.class,
-                      (message, context) -> targetsFrom(message.getId()))
+                      (message, context) -> targetsFrom(message.getEntity()))
                .route(EntityDeleted.class,
-                      (message, context) -> targetsFrom(message.getId()))
+                      (message, context) -> targetsFrom(message.getEntity()))
                .route(EntityUnarchived.class,
-                      (message, context) -> targetsFrom(message.getId()))
+                      (message, context) -> targetsFrom(message.getEntity()))
                .route(EntityRestored.class,
-                      (message, context) -> targetsFrom(message.getId()));
+                      (message, context) -> targetsFrom(message.getEntity()));
     }
 
-    private static Set<MirrorId> targetsFrom(EntityLogId historyId) {
+    private static Set<MirrorId> targetsFrom(MessageId historyId) {
         TypeUrl type = TypeUrl.parse(historyId.getTypeUrl());
         boolean shouldMirror = shouldMirror(type);
         return shouldMirror
@@ -128,13 +129,12 @@ final class MirrorRepository
         return visibility;
     }
 
-    private static MirrorId idFrom(EntityLogId historyId) {
-        Any any = historyId.getEntityId()
-                           .getId();
+    private static MirrorId idFrom(MessageId messageId) {
+        Any any = messageId.getId();
         MirrorId result = MirrorId
                 .newBuilder()
                 .setValue(any)
-                .setTypeUrl(historyId.getTypeUrl())
+                .setTypeUrl(messageId.getTypeUrl())
                 .build();
         return result;
     }
