+++ server/src/main/proto/spine/system/server/entity_log_events.proto
@@ -34,16 +34,16 @@ option java_multiple_files = true;
 import "google/protobuf/any.proto";
 import "google/protobuf/timestamp.proto";
 
+import "spine/core/diagnostics.proto";
 import "spine/core/command.proto";
 import "spine/core/event.proto";
 import "spine/core/version.proto";
-import "spine/system/server/entity_log.proto";
 
 // An entity was created.
 message EntityCreated {
 
     // The ID of the entity.
-    EntityLogId id = 1 [(required) = true];
+    core.MessageId entity = 1 [(required) = true];
 
     // The kind of the entity.
     EntityOption.Kind kind = 2 [(required) = true];
@@ -56,7 +56,7 @@ message EntityCreated {
 message EventDispatchedToSubscriber {
 
     // The ID of the entity, which the event is dispatched to.
-    EntityLogId receiver = 1 [(required) = true];
+    core.MessageId receiver = 1 [(required) = true];
 
     // The dispatched event.
     core.Event payload = 2 [(validate) = true];
@@ -76,7 +76,7 @@ message EventDispatchedToSubscriber {
 message EventDispatchedToReactor {
 
     // The ID of the entity, which the event is dispatched to.
-    EntityLogId receiver = 1 [(required) = true];
+    core.MessageId receiver = 1 [(required) = true];
 
     // The dispatched event.
     core.Event payload = 2 [(validate) = true];
@@ -96,7 +96,7 @@ message EventDispatchedToReactor {
 message CommandDispatchedToHandler {
 
     // The ID of the entity, which the command is dispatched to.
-    EntityLogId receiver = 1 [(required) = true];
+    core.MessageId receiver = 1 [(required) = true];
 
     // The dispatched command.
     core.Command payload = 2 [(validate) = true];
@@ -113,7 +113,7 @@ message CommandDispatchedToHandler {
 message EventImported {
 
     // The ID of the aggregate, which imported the event.
-    EntityLogId receiver = 1 [(required) = true];
+    core.MessageId receiver = 1 [(required) = true];
 
     // The info about the event.
     core.EventId event_id = 2 [(required) = true, (validate) = true];
@@ -129,13 +129,13 @@ message EventImported {
 message EntityStateChanged {
 
     // The ID of the entity.
-    EntityLogId id = 1 [(required) = true];
+    core.MessageId entity = 1 [(required) = true];
 
     // The new state of the entity.
     google.protobuf.Any new_state = 2 [(required) = true];
 
-    // The IDs of the messages which caused the mutation.
-    repeated DispatchedMessageId message_id = 3 [(required) = true, (validate) = true];
+    // The IDs of the signals which caused the mutation.
+    repeated core.MessageId signal_id = 3 [(required) = true, (validate) = true];
 
     // The instant when the entity was changed.
     google.protobuf.Timestamp when = 4;
@@ -148,10 +148,10 @@ message EntityStateChanged {
 message EntityArchived {
 
     // The ID of the entity.
-    EntityLogId id = 1 [(required) = true];
+    core.MessageId entity = 1 [(required) = true];
 
-    // The IDs of the messages which caused the mutation.
-    repeated DispatchedMessageId message_id = 2 [(required) = true, (validate) = true];
+    // The IDs of the signals which caused the mutation.
+    repeated core.MessageId signal_id = 2 [(required) = true, (validate) = true];
 
     // The instant when the entity was archived.
     google.protobuf.Timestamp when = 3;
@@ -164,10 +164,10 @@ message EntityArchived {
 message EntityDeleted {
 
     // The ID of the entity.
-    EntityLogId id = 1 [(required) = true];
+    core.MessageId entity = 1 [(required) = true];
 
-    // The IDs of the messages which caused the mutation.
-    repeated DispatchedMessageId message_id = 2 [(required) = true, (validate) = true];
+    // The IDs of the signals which caused the mutation.
+    repeated core.MessageId signal_id = 2 [(required) = true, (validate) = true];
 
     // The instant when the entity was deleted.
     google.protobuf.Timestamp when = 3;
@@ -180,10 +180,10 @@ message EntityDeleted {
 message EntityUnarchived {
 
     // The ID of the entity.
-    EntityLogId id = 1 [(required) = true];
+    core.MessageId entity = 1 [(required) = true];
 
-    // The IDs of the messages which caused the mutation.
-    repeated DispatchedMessageId message_id = 2 [(required) = true, (validate) = true];
+    // The IDs of the signals which caused the mutation.
+    repeated core.MessageId signal_id = 2 [(required) = true, (validate) = true];
 
     // The instant when the entity was extracted.
     google.protobuf.Timestamp when = 3;
@@ -196,10 +196,10 @@ message EntityUnarchived {
 message EntityRestored {
 
     // The ID of the entity.
-    EntityLogId id = 1 [(required) = true];
+    core.MessageId entity = 1 [(required) = true];
 
-    // The IDs of the messages which caused the mutation.
-    repeated DispatchedMessageId message_id = 2 [(required) = true, (validate) = true];
+    // The IDs of the signals which caused the mutation.
+    repeated core.MessageId signal_id = 2 [(required) = true, (validate) = true];
 
     // The instant when the entity was restored.
     google.protobuf.Timestamp when = 3;
