+++ server/src/test/java/io/spine/server/projection/ProjectionRepositoryTest.java
@@ -24,8 +24,8 @@ import com.google.common.collect.Lists;
 import com.google.protobuf.Any;
 import com.google.protobuf.Timestamp;
 import io.spine.base.EventMessage;
-import io.spine.client.EntityId;
 import io.spine.core.Event;
+import io.spine.core.MessageId;
 import io.spine.core.TenantId;
 import io.spine.core.Version;
 import io.spine.core.Versions;
@@ -44,7 +44,6 @@ import io.spine.server.type.EventClass;
 import io.spine.server.type.EventEnvelope;
 import io.spine.server.type.MessageEnvelope;
 import io.spine.server.type.given.GivenEvent;
-import io.spine.system.server.EntityLogId;
 import io.spine.system.server.event.EntityStateChanged;
 import io.spine.test.projection.Project;
 import io.spine.test.projection.ProjectId;
@@ -77,7 +76,7 @@ import static io.spine.base.Identifier.pack;
 import static io.spine.base.Time.currentTime;
 import static io.spine.server.projection.ProjectionRepository.nullToDefault;
 import static io.spine.server.projection.given.ProjectionRepositoryTestEnv.GivenEventMessage.projectCreated;
-import static io.spine.server.projection.given.ProjectionRepositoryTestEnv.dispatchedMessageId;
+import static io.spine.server.projection.given.ProjectionRepositoryTestEnv.dispatchedEventId;
 import static io.spine.testing.TestValues.randomString;
 import static io.spine.testing.server.Assertions.assertEventClasses;
 import static io.spine.testing.server.TestEventFactory.newInstance;
@@ -336,17 +335,17 @@ class ProjectionRepositoryTest
             dispatch(project, eventFactory.createEvent(projectCreated));
             dispatch(project, eventFactory.createEvent(taskAdded));
             Any newState = pack(project.state());
-            EntityLogId historyId = EntityLogId
+            MessageId entityId = MessageId
                     .newBuilder()
                     .setTypeUrl(newState.getTypeUrl())
-                    .setEntityId(EntityId.newBuilder().setId(pack(id)))
-                    .build();
+                    .setId(pack(id))
+                    .vBuild();
             EntityStateChanged changedEvent = EntityStateChanged
                     .newBuilder()
-                    .setId(historyId)
+                    .setEntity(entityId)
                     .setWhen(currentTime())
                     .setNewState(newState)
-                    .addMessageId(dispatchedMessageId())
+                    .addSignalId(dispatchedEventId())
                     .build();
             EntitySubscriberProjection.Repository repository =
                     new EntitySubscriberProjection.Repository();
